+++ pystac/serialization/__init__.py
@@ -1,3 +1,10 @@
+__all__ = [
+    "merge_common_properties",
+    "migrate_to_latest",
+    "STACVersionRange",
+    "identify_stac_object",
+    "identify_stac_object_type",
+]
 from pystac.serialization.identify import (
     STACVersionRange,
     identify_stac_object,