+++ tests/test_link.py
@@ -130,6 +130,73 @@ class LinkTest(unittest.TestCase):
             finally:
                 os.chdir(previous)
 
+    def test_auto_title_when_resolved(self) -> None:
+        extent = pystac.Extent.from_items([self.item])
+        collection = pystac.Collection(
+            id="my_collection",
+            description="Test Collection",
+            extent=extent,
+            title="Collection Title",
+        )
+        link = pystac.Link("my rel", target=collection)
+
+        self.assertEqual(collection.title, link.title)
+
+    def test_auto_title_not_found(self) -> None:
+        extent = pystac.Extent.from_items([self.item])
+        collection = pystac.Collection(
+            id="my_collection",
+            description="Test Collection",
+            extent=extent,
+        )
+        link = pystac.Link("my rel", target=collection)
+
+        self.assertEqual(None, link.title)
+
+    def test_auto_title_is_serialized(self) -> None:
+        extent = pystac.Extent.from_items([self.item])
+        collection = pystac.Collection(
+            id="my_collection",
+            description="Test Collection",
+            extent=extent,
+            title="Collection Title",
+        )
+        link = pystac.Link("my rel", target=collection)
+
+        assert link.to_dict().get("title") == collection.title
+
+    def test_no_auto_title_if_not_resolved(self) -> None:
+        link = pystac.Link(
+            "my rel", target="https://www.some-domain.com/path/to/thing.txt"
+        )
+
+        assert link.title is None
+
+    def test_title_as_init_argument(self) -> None:
+        link_title = "Link title"
+        extent = pystac.Extent.from_items([self.item])
+        collection = pystac.Collection(
+            id="my_collection",
+            description="Test Collection",
+            extent=extent,
+            title="Collection Title",
+        )
+        link = pystac.Link("my rel", title=link_title, target=collection)
+
+        assert link.title == link_title
+        assert link.to_dict().get("title") == link_title
+
+    def test_serialize_link(self) -> None:
+        href = "https://some-domain/path/to/item.json"
+        title = "A Test Link"
+        link = pystac.Link(pystac.RelType.SELF, href, pystac.MediaType.JSON, title)
+        link_dict = link.to_dict()
+
+        self.assertEqual(str(link_dict["rel"]), "self")
+        self.assertEqual(str(link_dict["type"]), "application/json")
+        self.assertEqual(link_dict["title"], title)
+        self.assertEqual(link_dict["href"], href)
+
 
 class StaticLinkTest(unittest.TestCase):
     def setUp(self) -> None: